Merge main

+++ pkg/gcmanager/tracePod_worker.go
@@ -11,8 +11,8 @@ import (
 	apierrors "k8s.io/apimachinery/pkg/api/errors"
 
 	"github.com/spidernet-io/spiderpool/pkg/constant"
-	metrics "github.com/spidernet-io/spiderpool/pkg/metric"
-	"github.com/spidernet-io/spiderpool/pkg/workloadendpointmanager"
+	"github.com/spidernet-io/spiderpool/pkg/metric"
+	"github.com/spidernet-io/spiderpool/pkg/utils/convert"
 )
 
 // tracePodWorker will circle traverse PodEntry database
@@ -69,7 +69,7 @@ func (s *SpiderGC) releaseIPPoolIPExecutor(ctx context.Context, workerIndex int)
 	for {
 		select {
 		case podCache := <-s.gcIPPoolIPSignal:
-			endpoint, err := s.wepMgr.GetEndpointByName(ctx, podCache.Namespace, podCache.PodName)
+			endpoint, err := s.wepMgr.GetEndpointByName(ctx, podCache.Namespace, podCache.PodName, constant.UseCache)
 			if nil != err {
 				if apierrors.IsNotFound(err) {
 					loggerReleaseIP.Sugar().Infof("SpiderEndpoint '%s/%s' not found, maybe already cleaned by ScanAll", podCache.Namespace, podCache.PodName)
@@ -80,9 +80,8 @@ func (s *SpiderGC) releaseIPPoolIPExecutor(ctx context.Context, workerIndex int)
 				continue
 			}
 
-			// we need to gather the pod corresponding SpiderEndpoint to get the used history IPs.
-			podUsedIPs := workloadendpointmanager.ListAllHistoricalIPs(endpoint)
-
+			// we need to gather the pod corresponding SpiderEndpoint allocation data to get the used history IPs.
+			podUsedIPs := convert.GroupIPAllocationDetails(endpoint.Status.Current.UID, endpoint.Status.Current.IPs)
 			// release pod used history IPs
 			for poolName, ips := range podUsedIPs {
 				loggerReleaseIP.Sugar().Infof("pod '%s/%s used IPs '%+v' from pool '%s', begin to release",
@@ -90,20 +89,18 @@ func (s *SpiderGC) releaseIPPoolIPExecutor(ctx context.Context, workerIndex int)
 
 				err = s.ippoolMgr.ReleaseIP(ctx, poolName, ips)
 				if nil != err {
-					metrics.IPGCFailureCounts.Add(ctx, 1)
-					loggerReleaseIP.Sugar().Errorf("failed to release pool '%s' IPs '%+v' in wep '%s/%s', error: %v",
+					metric.IPGCFailureCounts.Add(ctx, 1)
+					loggerReleaseIP.Sugar().Errorf("failed to release pool '%s' IPs '%+v' in SpiderEndpoint '%s/%s', error: %v",
 						poolName, ips, podCache.Namespace, podCache.PodName, err)
 					continue
 				}
-
-				// metric
-				metrics.IPGCTotalCounts.Add(ctx, 1)
+				metric.IPGCTotalCounts.Add(ctx, 1)
 			}
 
 			loggerReleaseIP.Sugar().Infof("release IPPoolIP task '%+v' successfully", *podCache)
 
 			// delete StatefulSet wep (other controller wep has OwnerReference, its lifecycle is same with pod)
-			if endpoint.Status.OwnerControllerType == constant.KindStatefulSet {
+			if endpoint.Status.OwnerControllerType == constant.KindStatefulSet && endpoint.DeletionTimestamp == nil {
 				err = s.wepMgr.DeleteEndpoint(ctx, endpoint)
 				if nil != err {
 					loggerReleaseIP.Sugar().Errorf("failed to delete StatefulSet wep '%s/%s', error: '%v'",
